extracted shows
