+++ projects/01_fyyur/starter_code/templates/layouts/main.html
@@ -80,7 +80,7 @@
           <ul class="nav navbar-nav">
             <li {% if request.endpoint == 'venues' %} class="active" {% endif %}><a href="{{ url_for('venue_api.venues') }}">Venues</a></li>
             <li {% if request.endpoint == 'artists' %} class="active" {% endif %}><a href="{{ url_for('artist_api.artists') }}">Artists</a></li>
-            <li {% if request.endpoint == 'shows' %} class="active" {% endif %}><a href="{{ url_for('shows') }}">Shows</a></li>
+            <li {% if request.endpoint == 'shows' %} class="active" {% endif %}><a href="{{ url_for('show_api.shows') }}">Shows</a></li>
           </ul>
         </div><!--/.nav-collapse -->
       </div>
